+++ internal/gengapic/gengrpc.go
@@ -185,6 +185,9 @@ func (g *generator) grpcClientOptions(serv *descriptorpb.ServiceDescriptorProto,
 	p("    internaloption.WithDefaultAudience(%q),", generateDefaultAudience(host))
 	p("    internaloption.WithDefaultScopes(DefaultAuthScopes()...),")
 	p("    internaloption.EnableJwtWithScope(),")
+	if _, ok := enableNewAuthLibraryBlocklist[g.serviceConfig.GetName()]; !ok {
+		p("internaloption.EnableNewAuthLibrary(),")
+	}
 	p("    option.WithGRPCDialOption(grpc.WithDefaultCallOptions(")
 	p("      grpc.MaxCallRecvMsgSize(math.MaxInt32))),")
 	p("  }")